implement model functions getQuestionsByProductId and getAnswersByQuestionId
