+++ models/index.js
@@ -1,0 +1,4 @@
+const question = require('./question');
+const answer = require('./answer');
+
+module.exports = { question, answer };
